separate out data-connector api
