+++ client/src/features/projectsV2/api/data-connectors.api.ts
@@ -1,4 +1,4 @@
-import { projectStoragesApi as api } from "./storagesV2.api";
+import { dataConnectorsEmptyApi as api } from "./data-connectors.empty-api";
 const injectedRtkApi = api.injectEndpoints({
   endpoints: (build) => ({
     getDataConnectors: build.query<
